src/requirements.txt
fastapi==0.115.11
langchain==0.3.20
langchain_community==0.3.19
langchain==0.3.20
langchain-core==0.3.43
langchain-text-splitters==0.3.6
pydantic==2.10.6
python-dotenv==1.0.1
faiss-cpu
openai
uvicorn
